all: add ts_omit_tailnetlock as a start of making it build-time modular
Updates #17115
